requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap4==2.2.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.8
psycopg2-binary==2.8.5
python-dotenv==0.14.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2
